refactor: extract update players status logic
